+++ NetCoreWebApiPoC.Application/NetCoreWebApiPoC.Application.csproj
@@ -20,8 +20,4 @@
     </Reference>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Context\" />
-  </ItemGroup>
-
 </Project>
