+++ NetCoreWebApiPoC.WebUI/NetCoreWebApiPoC.WebUI.csproj
@@ -24,7 +24,8 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\NetCoreWebApiPoC.Data\NetCoreWebApiPoC.Data.csproj" />
+    <ProjectReference Include="..\NetCoreWebApiPoC.Application\NetCoreWebApiPoC.Application.csproj" />
+    <ProjectReference Include="..\NetCoreWebApiPoC.Domain\NetCoreWebApiPoC.Domain.csproj" />
   </ItemGroup>
 
 </Project>
